Add junit reporter to the functests
Currently, the functests can be invoked with ginkgo's --junit-report
flag and produce a default junit report.

This commit adds a non-default junit report to the performance and
latency suites and configures it with specific configurations to produce
clean and easy-to-read report.

Added a "junit" flag, that when is not empty, will
produce the junit report in the given path.

Signed-off-by: liornoy <lnoy@redhat.com>